fix(config): properly set default macaroon paths
This commit fixes the derivation of the default lnd macaroon paths based
on the active network. Previously, changing the `network` config value
from `testnet` would not change the default macaroon paths, they would
need to be manually specified. This also correctly sets the testnet
directory for the litecoin network as `testnet4` instead of `testnet`.